reporting/streamlit/requirements.txt
streamlit 
matplotlib
google-cloud-secret-manager
numpy
pandas
duckdb==1.3.2

